transcribe/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
prisma==0.11.0
asyncpg==0.29.0

# Queue and Redis
redis==5.0.1
celery==5.3.4

# AWS and cloud services
boto3==1.34.0
botocore==1.34.0

# Transcription providers
assemblyai==0.20.0
openai==1.3.0

# Configuration and environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
httpx==0.25.2
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0